requirements.txt
fastapi
uvicorn
python-dotenv
google-generativeai
langchain
langchain-community
langchain-huggingface
langchain-chroma
sentence-transformers
chromadb
pydantic
torch




fastapi
uvicorn
python-dotenv
pydantic
google-generativeai
langchain
chromadb
langchain-huggingface
langchain-chroma
langchain-community
sentence-transformers
transformers
huggingface-hub
cross-encoder
torch
tqdm
playwright
python-multipart

